+++ packages/x-components/src/__stubs__/filters-stubs.factory.ts
@@ -1,12 +1,19 @@
 import {
-  HierarchicalFilter,
-  SimpleFilter,
-  NumberRangeFilter,
-  FilterModelName,
-  RangeValue,
   EditableNumberRangeFilter,
-  Filter
+  Filter,
+  FilterModelName,
+  HierarchicalFilter,
+  NumberRangeFilter,
+  RangeValue,
+  SimpleFilter
 } from '@empathyco/x-types';
+import {
+  EditableNumberRangeFilter as NextEditableNumberRangeFilter,
+  HierarchicalFilter as NextHierarchicalFilter,
+  NumberRangeFilter as NextNumberRangeFilter,
+  SimpleFilter as NextSimpleFilter,
+  RawFilter
+} from '@empathyco/x-types-next';
 
 /**
  * Creates {@link @empathyco/x-types#SimpleFilter | SimpleFilter} stub.
@@ -219,6 +226,7 @@ export function createSimpleFilter(facetId: string, label: string, selected = fa
  * and `label` fields.
  * @param selected - The selected value, false by default.
  * @returns An {@link @empathyco/x-types#HierarchicalFilter | HierarchicalFilter}.
+ * @deprecated Replace with {@link createNextHierarchicalFilter}.
  */
 export function createHierarchicalFilter(
   facetId: string,
@@ -255,5 +263,162 @@ export function createNumberRangeFilter(
     range,
     value: `{ "filter: "${min} to ${max}" }`,
     totalResults: 10
+  };
+}
+
+/* @empathyco/x-types-next factories */
+
+/**
+ * Creates a {@link @empathyco/x-types-next#RawFilter | RawFilter}.
+ *
+ * @param id - The identifier of the raw filter.
+ * @returns A {@link @empathyco/x-types-next#RawFilter | RawFilter} stub.
+ */
+export function createRawFilter(id: string): RawFilter {
+  return {
+    id,
+    modelName: 'RawFilter',
+    selected: true
+  };
+}
+
+/**
+ * Creates a {@link @empathyco/x-types-next#SimpleFilter | SimpleFilter}.
+ *
+ * @param facetId - The facet id this filter belongs to.
+ * @param label - The text that this filter should display.
+ * @param selected - True when the filter is checked, false otherwise.
+ * @returns A stub for a {@link @empathyco/x-types-next#SimpleFilter | SimpleFilter}.
+ */
+export function createNextSimpleFilter(
+  facetId: string,
+  label: string,
+  selected = false
+): NextSimpleFilter {
+  return {
+    id: `${facetId}:${label}`,
+    modelName: 'SimpleFilter',
+    facetId,
+    totalResults: 10,
+    label,
+    selected
+  };
+}
+
+/**
+ * Creates a {@link @empathyco/x-types-next#HierarchicalFilter | HierarchicalFilter}.
+ *
+ * @param facetId - The facet id this filter belongs to.
+ * @param label - The text that this filter should display.
+ * @param selected - True when the filter is checked, false otherwise.
+ * @param children - A list of ids of child filters.
+ * @returns A stub for a {@link @empathyco/x-types-next#HierarchicalFilter | HierarchicalFilter}.
+ */
+export function createNextHierarchicalFilter(
+  facetId: string,
+  label: string,
+  selected = false,
+  children: NextHierarchicalFilter['id'][] = []
+): NextHierarchicalFilter {
+  return {
+    facetId,
+    selected,
+    label,
+    id: `${facetId}:${label}`,
+    modelName: 'HierarchicalFilter',
+    parentId: null,
+    children
+  };
+}
+
+/**
+ * Creates a {@link @empathyco/x-types-next#NumberRangeFilter | NumberRangeFilter}.
+ *
+ * @param facetId - The facet id this filter belongs to.
+ * @param range - The range that this filter has.
+ * @param selected - True if the filter is selected, false otherwise.
+ * @returns A stub for a
+ * {@link @empathyco/x-types-next#NumberRangeFilter | NumberRangeFilter}.
+ */
+export function createNextNumberRangeFilter(
+  facetId: string,
+  range: RangeValue = { min: null, max: null },
+  selected = false
+): NextNumberRangeFilter {
+  return {
+    id: `${facetId}:${range.min ?? '*'}-${range.max ?? '*'}`,
+    modelName: 'NumberRangeFilter',
+    label: `From ${String(range.min)} to ${String(range.max)}`,
+    facetId,
+    range,
+    selected
+  };
+}
+
+/**
+ * Creates a {@link @empathyco/x-types-next#EditableNumberRangeFilter | EditableNumberRangeFilter}.
+ *
+ * @param facetId - The facet id this filter belongs to.
+ * @param range - The range that this filter has.
+ * @param selected - The selected value which has priority over the range values.
+ * @returns A stub for a
+ * {@link @empathyco/x-types-next#EditableNumberRangeFilter | EditableNumberRangeFilter}.
+ */
+export function createNextEditableNumberRangeFilter(
+  facetId: string,
+  range: RangeValue = { min: null, max: null },
+  selected?: boolean
+): NextEditableNumberRangeFilter {
+  return {
+    id: `${facetId}:${range.min ?? '*'}-${range.max ?? '*'}`,
+    facetId,
+    range,
+    modelName: 'EditableNumberRangeFilter',
+    selected: selected ?? (range.min !== null || range.max !== null)
+  };
+}
+
+/**
+ * Type of a function that creates
+ * {@link @empathyco/x-types-next#HierarchicalFilter | HierarchicalFilter}s. Based on its label,
+ * selected value and children.
+ */
+export type CreateNextHierarchicalFilter = (
+  label: string,
+  selected: boolean,
+  createChildren?: (createChildren: CreateNextHierarchicalFilter) => NextHierarchicalFilter[]
+) => NextHierarchicalFilter[];
+
+/**
+ * Creates a factory of
+ * {@link @empathyco/x-types-next#HierarchicalFilter | HierarchicalFilter}s.
+ *
+ * @remarks the id is created with `<facetId>:<label>`.
+ * @param facetId - Facet id to which the filter belongs.
+ * @param parentId - Filter's parent id if exists.
+ * @returns A scoped function which is able to create `HierarchicalFilters`.
+ */
+export function createNextHierarchicalFilterFactory(
+  facetId: string,
+  parentId: string | null = null
+): CreateNextHierarchicalFilter {
+  return (label, selected, createChildren): NextHierarchicalFilter[] => {
+    const filterId = `${facetId}:${label}`;
+    const children = createChildren
+      ? createChildren(createNextHierarchicalFilterFactory(facetId, filterId))
+      : [];
+    return [
+      {
+        id: filterId,
+        facetId: facetId,
+        parentId,
+        selected,
+        label,
+        totalResults: 10,
+        children: children.filter(filter => filter.parentId === filterId).map(filter => filter.id),
+        modelName: 'HierarchicalFilter'
+      },
+      ...children
+    ];
   };
 }