+++ packages/x-components/src/store/x.module.ts
@@ -31,6 +31,7 @@ export const RootXStoreModule: Module<XModuleState, any> = {
     identifierResults: null,
     search: null,
     facets: null,
+    facetsNext: null,
     tagging: null
   }),
   namespaced: true