+++ packages/x-components/src/x-modules/search/components/search-items-injection.mixin.ts
@@ -1,7 +1,7 @@
 import Vue from 'vue';
 import Component from 'vue-class-component';
 import { SearchItem } from '../../../utils/types';
-import { XInject } from '../../../components/decorators/injection.decorators';
+import { XInject, XProvide } from '../../../components/decorators/injection.decorators';
 import { SEARCH_ITEMS_KEY } from '../../../components/decorators/injection.consts';
 
 @Component
@@ -19,12 +19,15 @@ export default class SearchItemsInjectionMixin extends Vue {
   /**
    * The computed search items of the entity that uses the mixin.
    *
-   * @remarks It should be defined in the component that uses the mixin and it's intended to be
+   * @remarks It should be overridden in the component that uses the mixin and it's intended to be
    * filled with items from the state. Vue doesn't allow mixins as abstract classes.
-   *
+   * @returns An empty array as fallback in case it is not overridden.
    * @internal
    */
-  protected items!: SearchItem[];
+  @XProvide(SEARCH_ITEMS_KEY)
+  public get items(): SearchItem[] {
+    return [];
+  }
 
   /**
    * It injects {@link SearchItem} provided by an ancestor as injectedItems.